Create UI for Rock, Paper, Scissors game
Created DOM tree for UI. Added three buttons to select a choice. Added
function gameScore() and finalScore() to show a winner after 5 wins.
Added functions to disable buttons upon a win and re-enable buttons when
game is reset. Added function newButton() that resets the game when
there is a winner. Added function gameScore() to declare a winner. Added
some CSS to make it more presentable.